Refactor : userUtils 로 공통응답 처리 묶기
