+++ src/main/java/com/petmatz/domain/user/component/HeartService.java
@@ -3,31 +3,24 @@ package com.petmatz.domain.user.component;
 import com.petmatz.api.user.request.HeartedUserDto;
 import com.petmatz.api.user.request.HeartingRequestDto;
 import com.petmatz.common.security.utils.JwtExtractProvider;
-import com.petmatz.domain.user.entity.Heart;
 import com.petmatz.domain.user.entity.User;
-import com.petmatz.domain.user.entity.UserFactory;
 import com.petmatz.domain.user.repository.HeartRepository;
 import com.petmatz.domain.user.repository.UserRepository;
-import com.petmatz.domain.user.response.GetHeartingListResponseDto;
-import com.petmatz.domain.user.response.HeartingResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
 @Slf4j
 public class HeartService {
 
-    private final UserRepository userRepository;
     private final JwtExtractProvider jwtExtractProvider;
-    private final HeartRepository heartRepository;
     private final HeartComponent heartComponent;
+    private final UserUtils userUtils;
 
 
     @Transactional
@@ -36,7 +29,7 @@ public class HeartService {
         Long userId = jwtExtractProvider.findIdFromJwt();
 
         heartComponent.validateHeartUser(heartedId);
-        User currentUser = heartComponent.getCurrentUser(userId);
+        User currentUser = userUtils.getCurrentUser(userId);
         heartComponent.toggleHeart(currentUser.getId(), heartedId);
     }
